requirement.txt
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
loguru==0.6.0
pydantic==1.10.4
redis==4.5.1
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.25.0
typing_extensions==4.4.0
uvicorn==0.20.0
psycopg2-binary==2.9.2